[CHANGE] Load static files conditionally
