+++ tnnt_templates/static/tnnt_templates/javascript/terra-nanotech.js
@@ -1,8 +1,3 @@
-/* global CopyButtonPlugin */
-
-// Load the required libraries
-import hljs from '../libs/highlightjs/11.10.0/es/highlight.min.js';
-
 $(document).ready(() => {
     'use strict';
 
@@ -172,22 +167,6 @@ $(document).ready(() => {
         }, 500);
 
         externalLinks();
-
-        // Highlight code blocks with highlight.js
-        if (typeof hljs !== 'undefined' && typeof hljs.highlightElement === 'function') {
-            // Add the copy button plugin to highlight.js
-            hljs.addPlugin(new CopyButtonPlugin());
-
-            // Get all code blocks and highlight them
-            document.querySelectorAll('pre code[class*="language-"]').forEach((block) => {
-                const match = block.className.match(/language-(\w+)/);
-                const language = match ? match[1] : null;
-
-                if (language) {
-                    hljs.highlightElement(block, {language: language});
-                }
-            });
-        }
 
         /**
          * Set up overflow detection and monitoring for an element
