+++ tnnt_templates/static/tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-fittings.css
@@ -1,0 +1,197 @@
+@media all {
+    .aa-fittings #fitting-panel {
+        height: 398px;
+        margin: 0 auto;
+        width: 398px;
+        z-index: 3;
+    }
+
+    .aa-fittings #fitting-panel .fitting-item {
+        height: 32px;
+        width: 32px;
+    }
+
+    .aa-fittings #fitting-panel #mask > img {
+        height: 398px;
+        width: 398px;
+    }
+
+    .aa-fittings #fitting-panel #mask,
+    .aa-fittings #fitting-panel #highx,
+    .aa-fittings #fitting-panel #midx,
+    .aa-fittings #fitting-panel #lowx,
+    .aa-fittings #fitting-panel #rigx,
+    .aa-fittings #fitting-panel #subx {
+        height: 398px;
+        width: 398px;
+        z-index: -1;
+    }
+
+    .aa-fittings #fitting-panel #bigship {
+        height: 256px;
+        left: 72px;
+        top: 71px;
+        width: 256px;
+        z-index: -2;
+    }
+
+    .aa-fittings #fitting-panel #bigship > img {
+        height: 256px;
+        width: 256px;
+    }
+
+    .aa-fittings #fitting-panel #high1 {
+        left: 73px;
+        top: 60px;
+    }
+
+    .aa-fittings #fitting-panel #high2 {
+        left: 102px;
+        top: 42px;
+    }
+
+    .aa-fittings #fitting-panel #high3 {
+        left: 134px;
+        top: 27px;
+    }
+
+    .aa-fittings #fitting-panel #high4 {
+        left: 169px;
+        top: 21px;
+    }
+
+    .aa-fittings #fitting-panel #high5 {
+        left: 203px;
+        top: 22px;
+    }
+
+    .aa-fittings #fitting-panel #high6 {
+        left: 238px;
+        top: 30px;
+    }
+
+    .aa-fittings #fitting-panel #high7 {
+        left: 270px;
+        top: 45px;
+    }
+
+    .aa-fittings #fitting-panel #high8 {
+        left: 295px;
+        top: 64px;
+    }
+
+    .aa-fittings #fitting-panel #mid1 {
+        left: 26px;
+        top: 140px;
+    }
+
+    .aa-fittings #fitting-panel #mid2 {
+        left: 24px;
+        top: 176px;
+    }
+
+    .aa-fittings #fitting-panel #mid3 {
+        left: 23px;
+        top: 212px;
+    }
+
+    .aa-fittings #fitting-panel #mid4 {
+        left: 30px;
+        top: 245px;
+    }
+
+    .aa-fittings #fitting-panel #mid5 {
+        left: 46px;
+        top: 278px;
+    }
+
+    .aa-fittings #fitting-panel #mid6 {
+        left: 69px;
+        top: 304px;
+    }
+
+    .aa-fittings #fitting-panel #mid7 {
+        left: 100px;
+        top: 328px;
+    }
+
+    .aa-fittings #fitting-panel #mid8 {
+        left: 133px;
+        top: 342px;
+    }
+
+    .aa-fittings #fitting-panel #low1 {
+        left: 344px;
+        top: 143px;
+    }
+
+    .aa-fittings #fitting-panel #low2 {
+        left: 350px;
+        top: 178px;
+    }
+
+    .aa-fittings #fitting-panel #low3 {
+        left: 349px;
+        top: 213px;
+    }
+
+    .aa-fittings #fitting-panel #low4 {
+        left: 340px;
+        top: 246px;
+    }
+
+    .aa-fittings #fitting-panel #low5 {
+        left: 323px;
+        top: 277px;
+    }
+
+    .aa-fittings #fitting-panel #low6 {
+        left: 300px;
+        top: 304px;
+    }
+
+    .aa-fittings #fitting-panel #low7 {
+        left: 268px;
+        top: 324px;
+    }
+
+    .aa-fittings #fitting-panel #low8 {
+        left: 234px;
+        top: 338px;
+    }
+
+    .aa-fittings #fitting-panel #rig1 {
+        left: 148px;
+        top: 259px;
+    }
+
+    .aa-fittings #fitting-panel #rig2 {
+        left: 185px;
+        top: 267px;
+    }
+
+    .aa-fittings #fitting-panel #rig3 {
+        left: 221px;
+        top: 259px;
+    }
+
+    .aa-fittings #fitting-panel #sub1 {
+        left: 117px;
+        top: 131px;
+    }
+
+    .aa-fittings #fitting-panel #sub2 {
+        left: 147px;
+        top: 108px;
+    }
+
+    .aa-fittings #fitting-panel #sub3 {
+        left: 184px;
+        top: 98px;
+    }
+
+    .aa-fittings #fitting-panel #sub4 {
+        left: 221px;
+        top: 107px;
+    }
+}
