+++ tnnt_templates/static/tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-forum.css
@@ -1,0 +1,132 @@
+@media all {
+    .aa-forum .aa-forum-header {
+        margin-bottom: 1rem;
+    }
+
+    .aa-forum .aa-forum-admin-panel-aa-forum-category .panel-body {
+        min-height: auto;
+    }
+
+    .aa-forum .boards-sortable > li {
+        background-color: rgb(62 68 76 / 60%);
+    }
+
+    .aa-forum .panel-aa-forum-topic-list {
+        margin: 1.25rem 0;
+    }
+
+    .aa-forum .aa-forum-topic-row .aa-forum-topic:hover,
+    .aa-forum .panel-aa-forum-category .aa-forum-board:hover,
+    .aa-forum .panel-aa-forum-personal-messages-item:hover {
+        background-color: rgb(73 81 90 / 65%);
+    }
+
+    .aa-forum .aa-forum-topic-locked .aa-forum-topic:hover {
+        background-color: rgb(185 199 219 / 25%);
+    }
+
+    .aa-forum .aa-forum-topic-sticky .aa-forum-topic:hover {
+        background-color: rgb(255 220 104 / 25%);
+    }
+
+    .aa-forum .panel-aa-forum-personal-messages-item-header {
+        border-bottom: 1px solid rgb(200 200 200 / 25%);
+    }
+
+    .aa-forum .aa-forum-message-body,
+    .aa-forum .aa-forum-message-body-author-signature {
+        border-top: 1px solid rgb(200 200 200 / 25%);
+    }
+
+    .aa-forum .label-aa-forum-topic-new-message {
+        margin-left: 0.5rem;
+    }
+
+    .aa-forum .aa-forum-breadcrumb ul {
+        margin: 0;
+    }
+
+    .aa-forum .btn-aa-forum-topic-moderation {
+        border: none;
+    }
+
+    #aa-forum-form-search-menu .has-search svg {
+        fill: var(--font-color-default);
+    }
+
+    .aa-forum .aa-forum-search-results-total-number {
+        font-size: 1.25rem;
+        font-weight: bolder;
+        margin-bottom: 2rem;
+    }
+
+    .aa-forum .aa-forum-search-result-counter {
+        font-size: 2rem;
+        font-weight: bolder;
+        margin-right: 1rem;
+        min-width: 50px;
+        padding: 1rem;
+    }
+
+    .aa-forum #aa-forum-form-search-menu {
+        padding: 10px 0;
+    }
+
+    .aa-forum .aa-forum-breadcrumb {
+        padding: 0.2rem;
+    }
+
+    .aa-forum .aa-forum-legend > span {
+        margin-right: 3rem;
+    }
+
+    .aa-forum li.aa-forum-messages-sidebar-menu-item:has(a.active) {
+        background-color: var(--background-color-light-transparency) !important;
+    }
+
+    .aa-forum li.aa-forum-messages-sidebar-menu-item > a.active {
+        color: inherit !important;
+    }
+
+    /* SVG icons
+    --------------------------------------------------------------------------------- */
+
+    html[data-theme="terra-nanotech"] .aa-forum svg.svg-icon-post {
+        fill: rgb(133 133 133);
+    }
+
+    html[data-theme="terra-nanotech"] .aa-forum svg.svg-icon-post-new {
+        fill: rgb(255 255 255);
+    }
+
+    /* END SVG icons
+    --------------------------------------------------------------------------------- */
+}
+
+@media all and (min-width: 768px) {
+    .aa-forum #aa-forum-form-search-menu {
+        padding: 0;
+    }
+
+    .aa-forum .aa-forum-breadcrumb ul li:first-child {
+        padding-left: 0;
+    }
+}
+
+/* Highlight.js fixes
+------------------------------------------------------------------------------------- */
+
+@media all {
+    .hljs-copy-container {
+        top: 0.75rem;
+        right: 0.75rem;
+    }
+
+    .hljs-copy-button:not([data-copied="true"])::before {
+        width: 1.25rem;
+        height: 1.25rem;
+    }
+}
+
+/* END Highlight.js fixes
+------------------------------------------------------------------------------------- */
