+++ tnnt_templates/static/tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-sov-timer.css
@@ -1,0 +1,39 @@
+@media all {
+    .aa-sovtimer .aa-sovtimer-timer-elapsed {
+        color: rgb(253 126 20);
+        text-shadow: 1px 1px 1px rgb(255 0 0 / 30%);
+    }
+
+    .aa-sovtimer-active-campaign {
+        color: var(--bs-table-color);
+
+        --bs-table-active-bg: rgb(240 111 107);
+        --bs-table-bg: var(--bs-danger);
+        --bs-table-border-color: rgb(241 127 124);
+        --bs-table-hover-bg: rgb(239 107 103);
+        --bs-table-striped-bg: rgb(239 103 99);
+    }
+
+    .aa-sovtimer-active-campaign td {
+        background-color: rgb(238 95 91 / 25%) !important;
+    }
+
+    .aa-sovtimer-upcoming-campaign {
+        color: var(--bs-table-color);
+
+        --bs-table-active-bg: rgb(249 159 31);
+        --bs-table-bg: var(--bs-warning);
+        --bs-table-border-color: rgb(249 169 56);
+        --bs-table-hover-bg: rgb(249 156 25);
+        --bs-table-striped-bg: rgb(248 153 18);
+    }
+
+    .aa-sovtimer-upcoming-campaign td {
+        background-color: rgb(248 148 6 / 25%) !important;
+    }
+
+    .table-striped.aa-sovtimer-campaigns > tbody > tr:nth-of-type(odd) > *,
+    .table-striped.aa-sovtimer-campaigns > tbody > tr:nth-of-type(even) > * {
+        --bs-table-color-type: var(--bs-table-striped-color) !important;
+    }
+}
