+++ tnnt_templates/static/tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-srp.css
@@ -1,0 +1,7 @@
+@media all {
+    .aa-srp .card-srp-request-bulk-actions {
+        --background-color-light-transparency: rgb(58 63 68 / 95%);
+
+        background-color: var(--background-color-light-transparency) !important;
+    }
+}
