+++ tnnt_templates/static/tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/ckeditor-5.css
@@ -1,0 +1,26 @@
+@media all {
+    .ck-rounded-corners {
+        --ck-border-radius: var(--bs-border-radius);
+    }
+
+    .ck-editor__editable, .ck-editor__editable a {
+        color: initial;
+    }
+
+    .ck-word-count {
+        color: var(--bs-secondary-color) !important;
+
+        --bs-text-opacity: 1;
+    }
+
+    .ck-content blockquote {
+        border-left: 2px solid var(--bs-secondary-color) !important;
+        padding: 1rem 1rem 0;
+    }
+
+    .ck-content pre {
+        border: 1px solid var(--bs-secondary-color) !important;
+        border-radius: var(--bs-border-radius) !important;
+        padding: 1rem !important;
+    }
+}
