+++ tnnt_templates/templates/allianceauth/opengraph.html
@@ -1,5 +1,5 @@
 {% load evelinks %}
-{% load tnnt_template_tags %}
+{% load tnnt_templates %}
 
 <meta name="author" content="{{ TNNT_TEMPLATE_ENTITY_NAME }}">
