+++ tnnt_templates/templates/tnnt_templates/includes/theme/css.html
@@ -1,0 +1,35 @@
+{% load tnnt_templates %}
+{% load sri %}
+
+<!-- Terra Nanotech Theme CSS -->
+{% sri_static "tnnt_templates/theme/terra-nanotech/aav4/libs/bootstrap/v5.3.3/css/bootstrap.min.css" %}
+{% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/terra-nanotech-defaults.min.css" %}
+{% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/terra-nanotech-fonts.min.css" %}
+{% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/terra-nanotech.min.css" %}
+
+{% if request.path %}
+    {% if request.path|startswith:"/forum/" %}
+        <!-- AA Forum related CSS -->
+        {% sri_static "tnnt_templates/libs/fira-code/6.2.0/fira_code.min.css" %}
+        {% sri_static "tnnt_templates/libs/highlightjs/11.10.0/styles/github-dark.min.css" %}
+        {% sri_static "tnnt_templates/libs/highlightjs-copy/1.0.6/highlightjs-copy.min.css" %}
+        {% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-forum.min.css" %}
+        {% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/ckeditor-5.min.css" %}
+        <!-- AA Forum related CSS Ends -->
+    {% elif request.path|startswith:"/bulletin-board/" %}
+        {% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/ckeditor-5.min.css" %}
+    {% elif request.path|startswith:"/sovereignty-timer/" %}
+        <!-- AA Sovereignty Timer related CSS -->
+        {% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-sov-timer.min.css" %}
+        <!-- AA Sovereignty Timer related CSS Ends -->
+    {% elif request.path|startswith:"/fittings/" %}
+        <!-- AA Fittings related CSS -->
+        {% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-fittings.min.css" %}
+        <!-- AA Fittings related CSS Ends -->
+    {% elif request.path|startswith:"/ship-replacement/" %}
+        <!-- AA SRP related CSS -->
+        {% sri_static "tnnt_templates/theme/terra-nanotech/aav4/css/community-app-fixes/aa-srp.min.css" %}
+        <!-- AA SRP related CSS Ends -->
+    {% endif %}
+{% endif %}
+<!-- Terra Nanotech Theme CSS Ends -->
