+++ tnnt_templates/templates/tnnt_templates/includes/theme/js.html
@@ -1,0 +1,16 @@
+{% load tnnt_templates %}
+{% load sri %}
+
+<!-- Terra Nanotech Theme JS -->
+{% sri_static "tnnt_templates/theme/terra-nanotech/aav4/libs/popper/v2.11.8/popper.min.js" %}
+{% sri_static "tnnt_templates/theme/terra-nanotech/aav4/libs/bootstrap/v5.3.3/javascript/bootstrap.min.js" %}
+
+{% if request.path and request.path|startswith:"/forum/" %}
+    <!-- AA-Forum related JS -->
+    {% sri_static "tnnt_templates/libs/highlightjs-copy/1.0.6/highlightjs-copy.min.js" %}
+    {% sri_static "tnnt_templates/javascript/highlight.min.js" type="module" %}
+    <!-- AA-Forum related JS Ends -->
+{% endif %}
+
+{% sri_static "tnnt_templates/javascript/terra-nanotech.min.js" %}
+<!-- Terra Nanotech Theme JS Ends -->
